Cleanup: Expand notebook, migrate export process from graph class to export
