+++ src/spyglass/common/common_usage.py
@@ -6,11 +6,15 @@ determine which features are used, how often, and by whom. This will help
 plan future development of Spyglass.
 """
 
-from typing import Union
+from pathlib import Path
+from typing import List, Union
 
 import datajoint as dj
+from datajoint import FreeTable
+from datajoint import config as dj_config
 
 from spyglass.common.common_nwbfile import AnalysisNwbfile, Nwbfile
+from spyglass.settings import export_dir
 from spyglass.utils import SpyglassMixin, logger
 from spyglass.utils.dj_graph import RestrGraph
 from spyglass.utils.dj_helper_fn import unique_dicts
@@ -63,7 +67,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
         -> master
         table_id: int
         ---
-        table_name: varchar(64)
+        table_name: varchar(128)
         restriction: varchar(2048)
         """
 
@@ -82,6 +86,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
         """Custom insert to return export_id."""
         status = "Resuming"
         if not (query := self & key):
+            key = self._auto_increment(key, pk="export_id")
             super().insert1(key, **kwargs)
             status = "Starting"
         export_id = query.fetch1("export_id")
@@ -95,7 +100,7 @@ class ExportSelection(SpyglassMixin, dj.Manual):
         """Start logging a new export."""
         self._start_export(paper_id, analysis_id)
 
-    def stop_export(self) -> None:
+    def stop_export(self, **kwargs) -> None:
         """Stop logging the current export."""
         self._stop_export()
 
@@ -127,12 +132,12 @@ class ExportSelection(SpyglassMixin, dj.Manual):
         )
         return RestrGraph(seed_table=self, leaves=leaves, verbose=False)
 
-    def preview_tables(self, key: dict) -> list[dj.FreeTable]:
+    def preview_tables(self, **kwargs) -> list[dj.FreeTable]:
         """Return a list of restricted FreeTables for a given restriction/key.
 
         Useful for checking what will be exported.
         """
-        return self.get_restr_graph(key).leaf_ft
+        return self.get_restr_graph(kwargs).leaf_ft
 
     def _max_export_id(self, paper_id: str, return_all=False) -> int:
         """Return last export associated with a given paper id.
@@ -165,7 +170,7 @@ class Export(SpyglassMixin, dj.Computed):
         -> master
         table_id: int
         ---
-        table_name: varchar(64)
+        table_name: varchar(128)
         restriction: varchar(2048)
         unique index (table_name)
         """
@@ -222,8 +227,161 @@ class Export(SpyglassMixin, dj.Computed):
 
         # Writes but does not run mysqldump. Assumes single version per paper.
         version_key = query.fetch("spyglass_version", as_dict=True)[0]
-        restr_graph.write_export(**paper_key, **version_key)
+        self.write_export(
+            free_tables=restr_graph.all_ft, **paper_key, **version_key
+        )
 
         self.insert1(key)
-        self.Table().insert(table_inserts)  # TODO: Duplicate error??
+        self.Table().insert(table_inserts)
         self.File().insert(file_inserts)
+
+    def _get_credentials(self):
+        """Get credentials for database connection."""
+        return {
+            "user": dj_config["database.user"],
+            "password": dj_config["database.password"],
+            "host": dj_config["database.host"],
+        }
+
+    def _write_sql_cnf(self):
+        """Write SQL cnf file to avoid password prompt."""
+        cnf_path = Path("~/.my.cnf").expanduser()
+
+        if cnf_path.exists():
+            return
+
+        template = "[client]\nuser={user}\npassword={password}\nhost={host}\n"
+
+        with open(str(cnf_path), "w") as file:
+            file.write(template.format(**self._get_credentials()))
+        cnf_path.chmod(0o600)
+
+    def _bash_escape(self, s):
+        """Escape restriction string for bash."""
+        s = s.strip()
+
+        replace_map = {
+            "WHERE ": "",  # Remove preceding WHERE of dj.where_clause
+            "  ": " ",  # Squash double spaces
+            "( (": "((",  # Squash double parens
+            ") )": ")",
+            '"': "'",  # Replace double quotes with single
+            "`": "",  # Remove backticks
+            " AND ": " \\\n\tAND ",  # Add newline and tab for readability
+            " OR ": " \\\n\tOR  ",  # OR extra space to align with AND
+            ")AND(": ") \\\n\tAND (",
+            ")OR(": ") \\\n\tOR  (",
+            "#": "\\#",
+        }
+        for old, new in replace_map.items():
+            s = s.replace(old, new)
+        if s.startswith("(((") and s.endswith(")))"):
+            s = s[2:-2]  # Remove extra parens for readability
+        return s
+
+    def _cmd_prefix(self, docker_id=None):
+        """Get prefix for mysqldump command. Includes docker exec if needed."""
+        if not docker_id:
+            return "mysqldump "
+        return (
+            f"docker exec -i {docker_id} \\\n\tmysqldump "
+            + "-u {user} --password={password} \\\n\t".format(
+                **self._get_credentials()
+            )
+        )
+
+    def _write_mysqldump(
+        self,
+        free_tables: List[FreeTable],
+        paper_id: str,
+        docker_id=None,
+        spyglass_version=None,
+    ):
+        """Write mysqlmdump.sh script to export data.
+
+        Parameters
+        ----------
+        paper_id : str
+            Paper ID to use for export file names
+        docker_id : str, optional
+            Docker container ID to export from. Default None
+        spyglass_version : str, optional
+            Spyglass version to include in export. Default None
+        """
+        paper_dir = Path(export_dir) / paper_id if not docker_id else Path(".")
+        paper_dir.mkdir(exist_ok=True)
+
+        dump_script = paper_dir / f"_ExportSQL_{paper_id}.sh"
+        dump_content = paper_dir / f"_Populate_{paper_id}.sql"
+
+        prefix = self._cmd_prefix(docker_id)
+        version = (  # Include spyglass version as comment in dump
+            "echo '--'\n"
+            + f"echo '-- SPYGLASS VERSION: {spyglass_version} --'\n"
+            + "echo '--'\n\n"
+            if spyglass_version
+            else ""
+        )
+        create_cmd = (
+            "echo 'CREATE DATABASE IF NOT EXISTS {database}; "
+            + "USE {database};'\n\n"
+        )
+        dump_cmd = prefix + '{database} {table} --where="\\\n\t{where}"\n\n'
+
+        tables_by_db = sorted(free_tables, key=lambda x: x.full_table_name)
+
+        with open(dump_script, "w") as file:
+            file.write(
+                "#!/bin/bash\n\n"
+                + f"exec > {dump_content}\n\n"  # Redirect output to sql file
+                + f"{version}"  # Include spyglass version as comment
+            )
+
+            prev_db = None
+            for table in tables_by_db:
+                if not (where := table.where_clause()):
+                    continue
+                where = self._bash_escape(where)
+                database, table_name = (
+                    table.full_table_name.replace("`", "")
+                    .replace("#", "\\#")
+                    .split(".")
+                )
+                if database != prev_db:
+                    file.write(create_cmd.format(database=database))
+                    prev_db = database
+                file.write(
+                    dump_cmd.format(
+                        database=database, table=table_name, where=where
+                    )
+                )
+        logger.info(f"Export script written to {dump_script}")
+
+    def write_export(
+        self,
+        free_tables: List[FreeTable],
+        paper_id: str,
+        docker_id=None,
+        spyglass_version=None,
+    ):
+        """Write export bash script for all tables in graph.
+
+        Also writes a user-specific .my.cnf file to avoid password prompt.
+
+        Parameters
+        ----------
+        free_tables : List[FreeTable]
+            List of restricted FreeTables to export
+        paper_id : str
+            Paper ID to use for export file names
+        docker_id : str, optional
+            Docker container ID to export from. Default None
+        spyglass_version : str, optional
+            Spyglass version to include in export. Default None
+        """
+        self._write_sql_cnf()
+        self._write_mysqldump(
+            free_tables, paper_id, docker_id, spyglass_version
+        )
+
+        # TODO: export conda env
